Require parameters to TransmitMetadata::new()
The only time that values can be set for the underlying UHD C++
implementation of tx_metadata is at the time of creation. Therefore, we
must require that those values be provided to the constructor. Take in a
mandatory specification of burst start/stop markers and an optional
start time for the transmission.